Lexer
Got the lexer partially working. Reads in a file char by char, puts each
character into an array list. Reads the array list until a space (" ")
is found and then analyzes that chunk of characters as a string through
regular expressions. Adds them to a token array. Need to work on getting
this to loop correctly and add more regex