Tokens Complete
Got the tokens to add to the token array along with their value and
type. Accomplished this by using a counter which will run through the
regex, and count the number in the regex list where the match is found,
and then pass that through to the new token. The when a token is created
the createToken method is ran to see where the match took place, and
then assigns the token type accordingly. However, I still need to figure
out the line numbers.